+++ src/index.ts
@@ -1,9 +1,9 @@
-import * as club from './api/club';
-import * as competition from './api/competition';
-import * as country from './api/country';
-import * as matchday from './api/matchday';
-import * as player from './api/player';
-import * as season from './api/season';
+import * as club from './parsers/club';
+import * as competition from './parsers/competition';
+import * as country from './parsers/country';
+import * as matchday from './parsers/matchday';
+import * as player from './parsers/player';
+import * as season from './parsers/season';
 
 export { club, competition, country, matchday, player, season };
 export default { club, competition, country, matchday, player, season };
