+++ src/parsers/competition/interface.ts
