+++ src/parsers/matchday/__tests__/methods.test.ts
@@ -2,7 +2,7 @@ import { list } from '../methods';
 
 import { JEST_ASYNC_TIMEOUT } from '../../../settings';
 
-describe('src/api/matchday/methods.ts', () => {
+describe('src/parsers/matchday/methods.ts', () => {
     describe('list', () => {
         it(
             'Should return correct result for EPL, 17/18 (GB1, 2017)',
