+++ src/parsers/matchday/methods.ts
@@ -1,0 +1,25 @@
+import { JSDOM } from 'jsdom';
+
+import { Matchday } from './interface';
+
+// Utils
+import { makeRequest, parse } from '../../utils';
+
+import url from '../../url';
+
+export function list(competitionId: string, seasonId: string): Promise<Array<Matchday>> {
+    return makeRequest(url.matchday.list(competitionId, seasonId)).then(
+        parse(data => {
+            const dom = new JSDOM(data);
+
+            return [...dom.window.document.querySelectorAll('.large-6.columns:not([id])')]
+                .filter(node => !!node.querySelector('.table-header'))
+                .map((node, index) => ({
+                    competitionId,
+                    id: index + 1,
+                    seasonId,
+                    title: node.querySelector('.table-header').innerHTML,
+                }));
+        }),
+    );
+}
