+++ src/parsers/season/methods.ts
@@ -1,0 +1,24 @@
+import { JSDOM } from 'jsdom';
+
+import { Season } from './interface';
+
+// Utils
+import { makeRequest, parse } from '../../utils';
+
+import url from '../../url';
+
+export function list(competitionId: string): Promise<Array<Season>> {
+    return makeRequest(url.season.list(competitionId)).then(
+        parse(data => {
+            const dom = new JSDOM(data);
+
+            return [...dom.window.document.querySelectorAll('select[name=saison_id] option')]
+                .filter(node => node.getAttribute('value'))
+                .map(node => ({
+                    competitionId,
+                    id: node.getAttribute('value'),
+                    title: node.innerHTML,
+                }));
+        }),
+    );
+}
