+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+import makeRequest from './make-request';
+import parse from './parse';
+
+export { makeRequest, parse };
+export default { makeRequest, parse };
