+++ src/utils/parse.ts
@@ -1,0 +1,10 @@
+// Constants
+import { ERROR_PARSING } from '../constants/errors';
+
+export default fn => data => {
+    try {
+        return fn(data);
+    } catch (error) {
+        throw ERROR_PARSING;
+    }
+};
